Move JRubyPluginTest into Spock
This was just some tech debt lying around

Fixes #148
